+++ test/chart/data_pool.dart
@@ -1186,9 +1186,17 @@ List<LineTooltipItem?> lineChartGetTooltipItems(List<LineBarSpot> list) {
   return list.map((s) => lineTooltipItem1).toList();
 }
 
+Color lineChartGetGreenColor(LineBarSpot touchedSpot) {
+  return Colors.green;
+}
+
+Color lineChartGetRedColor(LineBarSpot touchedSpot) {
+  return Colors.red;
+}
+
 const LineTouchTooltipData lineTouchTooltipData1 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1198,7 +1206,7 @@ const LineTouchTooltipData lineTouchTooltipData1 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData1Clone = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1209,7 +1217,7 @@ const LineTouchTooltipData lineTouchTooltipData1Clone = LineTouchTooltipData(
 
 const LineTouchTooltipData lineTouchTooltipData2 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.red,
+  getTooltipColor: lineChartGetRedColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1219,7 +1227,7 @@ const LineTouchTooltipData lineTouchTooltipData2 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData3 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.2),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1230,7 +1238,7 @@ const LineTouchTooltipData lineTouchTooltipData3 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData4 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 13,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1241,7 +1249,7 @@ const LineTouchTooltipData lineTouchTooltipData4 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData5 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1252,7 +1260,7 @@ const LineTouchTooltipData lineTouchTooltipData5 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData6 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -1264,7 +1272,7 @@ const LineTouchTooltipData lineTouchTooltipData6 = LineTouchTooltipData(
 );
 const LineTouchTooltipData lineTouchTooltipData7 = LineTouchTooltipData(
   tooltipPadding: EdgeInsets.all(0.1),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: lineChartGetGreenColor,
   maxContentWidth: 12,
   getTooltipItems: lineChartGetTooltipItems,
   fitInsideHorizontally: true,
@@ -2240,6 +2248,14 @@ ScatterTooltipItem? scatterChartGetTooltipItems(ScatterSpot spots) {
   );
 }
 
+Color scatterChartGetTooltipGreenColor(ScatterSpot spots) {
+  return Colors.green; //Color
+}
+
+Color scatterChartGetTooltipRedColor(ScatterSpot spots) {
+  return Colors.red; //Color
+}
+
 final ScatterSpot scatterSpot1 = ScatterSpot(1, 40);
 final ScatterSpot scatterSpot1Clone = ScatterSpot(1, 40);
 final ScatterSpot scatterSpot2 = ScatterSpot(-4, -8);
@@ -2302,7 +2318,7 @@ final ScatterChartData scatterChartData1 = ScatterChartData(
       fitInsideHorizontally: true,
       fitInsideVertically: false,
       maxContentWidth: 33,
-      tooltipBgColor: Colors.white,
+      getTooltipColor: (touchedSpot) => Colors.white,
       tooltipPadding: const EdgeInsets.all(23),
       tooltipRoundedRadius: 534,
     ),
@@ -2341,7 +2357,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData1 =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2352,7 +2368,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData1Clone =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2363,7 +2379,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData2 =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2375,7 +2391,7 @@ final ScatterTouchTooltipData scatterTouchTooltipData3 =
     ScatterTouchTooltipData(
   tooltipRoundedRadius: 23,
   tooltipPadding: const EdgeInsets.all(11),
-  tooltipBgColor: Colors.green,
+  getTooltipColor: scatterChartGetTooltipGreenColor,
   maxContentWidth: 33,
   fitInsideVertically: true,
   fitInsideHorizontally: false,
@@ -2749,12 +2765,24 @@ BarTooltipItem getTooltipItem(
   return BarTooltipItem(rod.toY.toString(), textStyle);
 }
 
+Color getTooltipGreenColor(
+  BarChartGroupData group,
+) {
+  return Colors.green;
+}
+
+Color getTooltipBlueColor(
+  BarChartGroupData group,
+) {
+  return Colors.blue;
+}
+
 final BarTouchTooltipData barTouchTooltipData1 = BarTouchTooltipData(
   tooltipRoundedRadius: 12,
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2765,7 +2793,7 @@ final BarTouchTooltipData barTouchTooltipData1Clone = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2776,7 +2804,7 @@ final BarTouchTooltipData barTouchTooltipData2 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2788,7 +2816,7 @@ final BarTouchTooltipData barTouchTooltipData3 = BarTouchTooltipData(
   fitInsideVertically: true,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2800,7 +2828,7 @@ final BarTouchTooltipData barTouchTooltipData4 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: false,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2812,7 +2840,7 @@ final BarTouchTooltipData barTouchTooltipData5 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23.00001,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2825,7 +2853,7 @@ final BarTouchTooltipData barTouchTooltipData6 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.blue,
+  getTooltipColor: getTooltipBlueColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2838,7 +2866,7 @@ final BarTouchTooltipData barTouchTooltipData7 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
   tooltipBorder: const BorderSide(color: Colors.red),
@@ -2850,7 +2878,7 @@ final BarTouchTooltipData barTouchTooltipData8 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   tooltipMargin: 12,
   tooltipBorder: const BorderSide(color: Colors.red),
@@ -2860,7 +2888,7 @@ final BarTouchTooltipData barTouchTooltipData9 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 333,
@@ -2871,7 +2899,7 @@ final BarTouchTooltipData barTouchTooltipData10 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,
@@ -2882,7 +2910,7 @@ final BarTouchTooltipData barTouchTooltipData11 = BarTouchTooltipData(
   fitInsideVertically: false,
   fitInsideHorizontally: true,
   maxContentWidth: 23,
-  tooltipBgColor: Colors.green,
+  getTooltipColor: getTooltipGreenColor,
   tooltipPadding: const EdgeInsets.all(23),
   getTooltipItem: getTooltipItem,
   tooltipMargin: 12,